Another checkpoint

+++ src/com/dmdev/http/entity/Flight.java
@@ -10,13 +10,13 @@ public class Flight {
     private String departureAirportCode;
     private LocalDateTime arrivalDate;
     private String arrivalAirportCode;
-    private Long aircraftId;
+    private Integer aircraftId;
     private FlightStatus status;
 
     public Flight(Long id, String flightNo,
                   LocalDateTime departureDate, String departureAirportCode,
                   LocalDateTime arrivalDate, String arrivalAirportCode,
-                  Long aircraftId, FlightStatus status) {
+                  Integer aircraftId, FlightStatus status) {
         this.id = id;
         this.flightNo = flightNo;
         this.departureDate = departureDate;
@@ -75,11 +75,11 @@ public class Flight {
         this.arrivalAirportCode = arrivalAirportCode;
     }
 
-    public Long getAircraftId() {
+    public Integer getAircraftId() {
         return aircraftId;
     }
 
-    public void setAircraftId(Long aircraftId) {
+    public void setAircraftId(Integer aircraftId) {
         this.aircraftId = aircraftId;
     }
